A neural network with a single hidden layer
Get dataset called “Fashion-MNIST”. This dataset contains 10 different categories of clothing.
this neural network classifies this data.
Data. Each image is 28 pixels in height and 28 pixels in width, for a total
of 784 pixels in total. Each pixel has a single pixel-value associated with
it, indicating the lightness or darkness of that pixel. This pixel-value is an
integer between 0 and 255.
10 Labels - top, trouser, pullover, dress, coat, sandal,shrit,bag,sneaker,ankle boot
My neural network uses sigmoid and sofMax.
Note - I trained the network on a 55,000 sample set, the file size is too large so I created a smaller sample file.